+++ src/components/Navigation/Sidebar/Sidebar.js
@@ -1,6 +1,16 @@
 import { SideMenu, Logo, NavMenu, MenuItem, SocialLinksMenu, PullTab, LegalMenu } from "./Sidebar.styles";
 import { NavLink } from "react-router-dom";
-import { FaChartLine, FaArrowUp, FaShoppingCart, FaAward, FaBook, FaCoins, FaFile, FaLayerGroup } from "react-icons/fa";
+import {
+  FaChartLine,
+  FaArrowUp,
+  FaShoppingCart,
+  FaAward,
+  FaBook,
+  FaCoins,
+  FaFile,
+  FaLayerGroup,
+  FaSync,
+} from "react-icons/fa";
 
 import logoImg from "../../../img/logo_MYC.svg";
 import gitbookIcon from "../../../img/gitbook.svg";
@@ -34,6 +44,11 @@ const navLinks = [
     name: "Rewards",
     path: "/rewards",
     icon: FaAward,
+  },
+  {
+    name: "Referrals",
+    path: "/referrals",
+    icon: FaSync,
   },
   {
     name: "Docs",
